remove some tests from TriggerTest
