+++ full/src/test/java/apoc/trigger/TriggerExtendedTest.java
@@ -6,10 +6,14 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.neo4j.graphdb.Label;
 import org.neo4j.graphdb.Node;
+import org.neo4j.graphdb.QueryExecutionException;
 import org.neo4j.test.rule.DbmsRule;
 import org.neo4j.test.rule.ImpermanentDbmsRule;
 
+import java.util.Collections;
 import java.util.Map;
+import java.util.concurrent.Callable;
+import java.util.concurrent.TimeUnit;
 
 import static apoc.ApocSettings.apoc_trigger_enabled;
 import static org.junit.Assert.assertEquals;
@@ -38,7 +42,7 @@ public class TriggerExtendedTest {
         db.executeTransactionally("CALL apoc.trigger.add('timestamp','UNWIND apoc.trigger.nodesByLabel($assignedNodeProperties,null) AS n SET n.ts = timestamp()',{})");
         db.executeTransactionally("CREATE (f:Foo) SET f.foo='bar'");
         TestUtil.testCall(db, "MATCH (f:Foo) RETURN f", (row) -> {
-            assertEquals(true, ((Node)row.get("f")).hasProperty("ts"));
+            assertEquals(true, ((Node) row.get("f")).hasProperty("ts"));
         });
     }
 
@@ -48,22 +52,42 @@ public class TriggerExtendedTest {
         db.executeTransactionally("CALL apoc.trigger.add('lowercase','UNWIND apoc.trigger.nodesByLabel($assignedLabels,\"Person\") AS n SET n.id = toLower(n.name)',{})");
         db.executeTransactionally("CREATE (f:Person {name:'John Doe'})");
         TestUtil.testCall(db, "MATCH (f:Person) RETURN f", (row) -> {
-            assertEquals("john doe", ((Node)row.get("f")).getProperty("id"));
-            assertEquals("John Doe", ((Node)row.get("f")).getProperty("name"));
+            assertEquals("john doe", ((Node) row.get("f")).getProperty("id"));
+            assertEquals("John Doe", ((Node) row.get("f")).getProperty("name"));
         });
     }
+
     @Test
     public void testSetLabels() throws Exception {
         db.executeTransactionally("CREATE (f {name:'John Doe'})");
         db.executeTransactionally("CALL apoc.trigger.add('setlabels','UNWIND apoc.trigger.nodesByLabel($assignedLabels,\"Person\") AS n SET n:Man',{})");
         db.executeTransactionally("MATCH (f) SET f:Person");
         TestUtil.testCall(db, "MATCH (f:Man) RETURN f", (row) -> {
-            assertEquals("John Doe", ((Node)row.get("f")).getProperty("name"));
-            assertEquals(true, ((Node)row.get("f")).hasLabel(Label.label("Person")));
+            assertEquals("John Doe", ((Node) row.get("f")).getProperty("name"));
+            assertEquals(true, ((Node) row.get("f")).hasLabel(Label.label("Person")));
         });
 
         long count = TestUtil.singleResultFirstColumn(db, "MATCH (f:Man) RETURN count(*) as c");
         assertEquals(1L, count);
+    }
+
+    @Test
+    public void testTxIdAfterAsync() throws Exception {
+        db.executeTransactionally("CALL apoc.trigger.add('triggerTest','UNWIND apoc.trigger.propertiesByKey($assignedNodeProperties, \"_executed\") as prop " +
+                "	WITH prop.node as n " +
+                "	CREATE (z:SON {father:id(n)}) " +
+                "	CREATE (n)-[:GENERATED]->(z)', " +
+                "{phase:'afterAsync'})");
+        db.executeTransactionally("CREATE (:TEST {name:'x', _executed:0})");
+        db.executeTransactionally("CREATE (:TEST {name:'y', _executed:0})");
+        org.neo4j.test.assertion.Assert.assertEventually((Callable<Long>) () -> db.executeTransactionally("MATCH p = ()-[r:GENERATED]->() RETURN count(p) AS count",
+                Collections.emptyMap(), (r) -> r.<Long>columnAs("count").next()),
+                value -> value.equals(2L), 30L, TimeUnit.SECONDS);
+    }
+
+    @Test(expected = QueryExecutionException.class)
+    public void showThrowAnException() throws Exception {
+        db.executeTransactionally("CALL apoc.trigger.add('test','UNWIND $createdNodes AS n SET n.txId = , n.txTime = $commitTime',{})");
     }
 
 }
